oraculo_agente_ia/requirements.txt
# ==========================================
# API / Runtime
# ==========================================
fastapi==0.135.3
uvicorn==0.44.0
httptools==0.7.1
watchfiles==1.1.1
websockets==16.0
sse-starlette==3.3.4
python-multipart==0.0.20

# ==========================================
# Configuration / Security
# ==========================================
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
PyJWT==2.12.1
httpx==0.28.1

# ==========================================
# Database / Storage
# ==========================================
SQLAlchemy==2.0.49
aiosqlite==0.21.0
qdrant-client==1.17.1

# ==========================================
# LangChain Ecosystem
# ==========================================
langchain==1.2.15
langgraph==1.1.6
langgraph-checkpoint-sqlite==3.0.3
langchain-google-genai==4.2.1
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langserve==0.3.3
langmem==0.0.30
langsmith==0.4.46

# ==========================================
# Testing / Quality
# ==========================================
pytest==9.0.3
pytest-asyncio==1.3.0
hypothesis==6.151.12
schemathesis==4.15.1
